renaming `createLambdaJSON` to `createFunction`, move to export on module, lowercase and dasherize create-function module
